requirements.txt
# This file is for development and testing only.
# Runtime requirements are in setup.py install_requirements.

factory_boy
pytest
pytest-django
pytest-mock
pytest-pythonpath
